chore: address cargo clippy issues (#6318)

+++ lib/napi/src/guest/util.rs
@@ -80,7 +80,7 @@ pub fn host_ptr_to_guest_ptr(env: &mut FunctionEnvMut<RuntimeEnv>, host_addr: u6
     let (_, store_ref) = env.data_and_store_mut();
     let view = memory.view(&store_ref);
     let host_base = view.data_ptr() as u64;
-    let memory_len = view.data_size() as u64;
+    let memory_len = view.data_size();
     if host_addr < host_base || host_addr >= host_base + memory_len {
         return None;
     }
